feat(create-account): implement create account logic
- Add custom hook for create account logic
- Integrate form handling with validation
- Manage secure text entry state
- Implement navigation to login page after account creation
